requirements.txt
pandas
pyyaml
jinja2
python-docx
openai>=1.14
openpyxl
docxtpl